Fix compile error with gcc

+++ lib/ngtcp2_pkt.c
@@ -2257,7 +2257,7 @@ size_t ngtcp2_pkt_crypto_max_datalen(uint64_t offset, size_t len, size_t left) {
 }
 
 uint8_t ngtcp2_pkt_get_type_long(uint8_t c) {
-  return (c & NGTCP2_LONG_TYPE_MASK) >> 4;
+  return (uint8_t)((c & NGTCP2_LONG_TYPE_MASK) >> 4);
 }
 
 int ngtcp2_pkt_verify_reserved_bits(uint8_t c) {
